Changed background drawing functions to use float precision.

+++ ENIGMAsystem/SHELL/Graphics_Systems/General/GSbackground.h
@@ -38,7 +38,7 @@ double background_get_texture_height_factor(int backId);
 @param x
 @param y
 **/
-void draw_background(int back, double x, double y);
+void draw_background(int back, float x, float y);
 /** description
 @param back The index of the background asset.
 @param x
@@ -56,13 +56,13 @@ void draw_background_stretched(int back, double x, double y, double w, double h)
 @param x
 @param y
 **/
-void draw_background_part(int back,double left,double top,double width,double height,double x,double y);
+void draw_background_part(int back, float left, float top, float width, float height, float x, float y);
 /** description
 @param back The index of the background asset.
 @param x
 @param y
 **/
-void draw_background_tiled(int back,double x,double y);
+void draw_background_tiled(int back, float x, float y);
 /** description
 @param back The index of the background asset.
 @param x
@@ -72,7 +72,7 @@ void draw_background_tiled(int back,double x,double y);
 @param x2
 @param y2
 **/
-void draw_background_tiled_area(int back,double x,double y,double x1,double y1,double x2,double y2);
+void draw_background_tiled_area(int back,float x,float y,float x1,float y1,float x2,float y2);
 /** description
 @param back The index of the background asset.
 @param x
@@ -83,7 +83,7 @@ void draw_background_tiled_area(int back,double x,double y,double x1,double y1,d
 @param color
 @param alpha
 **/
-void draw_background_ext(int back,double x,double y,double xscale,double yscale,double rot,int color,double alpha);
+void draw_background_ext(int back, float x, float y, float xscale, float yscale, double rot, int color, double alpha);
 /** description
 @param back The index of the background asset.
 @param x
@@ -93,7 +93,7 @@ void draw_background_ext(int back,double x,double y,double xscale,double yscale,
 @param color
 @param alpha
 **/
-void draw_background_stretched_ext(int back,double x,double y,double w,double h,int color,double alpha);
+void draw_background_stretched_ext(int back, float x, float y, float w, float h, int color, double alpha);
 /** description
 @param back The index of the background asset.
 @param left
@@ -107,7 +107,7 @@ void draw_background_stretched_ext(int back,double x,double y,double w,double h,
 @param color
 @param alpha
 **/
-void draw_background_part_ext(int back,double left,double top,double width,double height,double x,double y,double xscale,double yscale,int color,double alpha);
+void draw_background_part_ext(int back, float left, float top, float width, float height, float x, float y, float xscale, float yscale,int color, double alpha);
 /** description
 @param back The index of the background asset.
 @param x
@@ -117,7 +117,7 @@ void draw_background_part_ext(int back,double left,double top,double width,doubl
 @param color
 @param alpha
 **/
-void draw_background_tiled_ext(int back,double x,double y,double xscale,double yscale,int color,double alpha);
+void draw_background_tiled_ext(int back, float x, float y, float xscale, float yscale, int color, double alpha);
 /** description
 @param back The index of the background asset.
 @param x
@@ -131,7 +131,7 @@ void draw_background_tiled_ext(int back,double x,double y,double xscale,double y
 @param color
 @param alpha
 **/
-void draw_background_tiled_area_ext(int back,double x,double y,double x1,double y1,double x2,double y2, double xscale, double yscale, int color, double alpha);
+void draw_background_tiled_area_ext(int back, float x, float y, float x1, float y1, float x2, float y2, float xscale, float yscale, int color, double alpha);
 /** description
 @param back The index of the background asset.
 @param left
@@ -152,7 +152,7 @@ void draw_background_tiled_area_ext(int back,double x,double y,double x1,double 
 @param a3
 @param a4
 **/
-void draw_background_general(int back,double left,double top,double width,double height,double x,double y,double xscale,double yscale,double rot,int c1,int c2,int c3,int c4,double a1,double a2,double a3,double a4);
+void draw_background_general(int back, float left, float top, float width, float height, float x, float y, float xscale, float yscale, double rot, int c1, int c2, int c3, int c4, double a1, double a2, double a3, double a4);
 
 }
